nsfw_13b_sft_gguf/causal_lm/pytorch/requirements.txt
gguf>=0.10.0
transformers_stream_generator
